tests/test_kosu.py
from pathlib import Path

from click.testing import CliRunner

from kosu import cli


def test_help():
    """
    Test the help shows up.
    """
    runner = CliRunner()
    result = runner.invoke(cli, ['--help'])
    assert result.exit_code == 0
    assert 'Show this message and exit.' in result.output


def test_missing():
    """
    Test the app does the right thing with a missing arg.
    """
    runner = CliRunner()
    result = runner.invoke(cli, ['test'])
    assert result.exit_code == 2
    assert "Error: Missing argument 'COURSE'" in result.output


def test_clean():
    """
    Test the clean function runs and removes the build folder.
    """
    runner = CliRunner()
    with runner.isolated_filesystem():
        _ = Path('build').mkdir()
        result = runner.invoke(cli, ['clean', '--all'])
        assert result.exit_code == 0
        assert "Finished." in result.output
        assert not Path('build').is_file()


def test_init():
    """
    Test that project initialization works.
    """
    runner = CliRunner()
    with runner.isolated_filesystem():
        result = runner.invoke(cli, ['init', '--yes'])
        # These seem to cause problems, should be enough to check for YAML file.
        # assert result.exit_code == 0
        # assert 'Created example course' in result.output
        assert Path('example_course.yaml').is_file()
